refactor: 스케쥴러, 기타 기능 리팩토링
