+++ src/main/java/hhplus/booking/app/user/infra/jpa/UserJpaRepository.java
@@ -1,10 +1,15 @@
 package hhplus.booking.app.user.infra.jpa;
 
 import hhplus.booking.app.user.domain.entity.User;
-import hhplus.booking.app.user.domain.repository.UserRepository;
+import jakarta.persistence.LockModeType;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Lock;
 import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
 
 @Repository
 public interface UserJpaRepository extends JpaRepository<User, Long> {
+    @Lock(LockModeType.PESSIMISTIC_WRITE)
+    Optional<User> findById(Long id);
 }
